requirements.txt
# for selenium
requests==2.25.1
requests-oauthlib==1.3.0
PyYAML==5.4.1
